[SYSTEMML-672] Fix PyDML 1-Based Indexing Bug
Currently, PyDML has a *bug* in which it uses 1-based indexing, as in R & DML, rather than the correct 0-based indexing that Python uses.  This simply aims to fix this issue in the Parser, and then update the few PyDML test scripts to use the correct 0-based indexing.
