+++ src/main/java/org/apache/sysml/parser/common/CommonSyntacticValidator.java
@@ -379,70 +379,6 @@ public abstract class CommonSyntacticValidator {
 		}
 	}
 
-	protected void exitIndexedExpressionHelper(ParserRuleContext ctx, String name, ExpressionInfo dataInfo,
-			ExpressionInfo rowLower, ExpressionInfo rowUpper, ExpressionInfo colLower, ExpressionInfo colUpper) {
-		dataInfo.expr = new IndexedIdentifier(name, false, false);
-		setFileLineColumn(dataInfo.expr, ctx);
-		boolean isRowLower = rowLower != null;
-		boolean isRowUpper = rowUpper != null;
-		boolean isColLower = colLower != null;
-		boolean isColUpper = colUpper != null;
-		try {
-			ArrayList< ArrayList<Expression> > exprList = new ArrayList< ArrayList<Expression> >();
-
-			ArrayList<Expression> rowIndices = new ArrayList<Expression>();
-			ArrayList<Expression> colIndices = new ArrayList<Expression>();
-
-
-			if(!isRowLower && !isRowUpper) {
-				// both not set
-				rowIndices.add(null); rowIndices.add(null);
-			}
-			else if(isRowLower && isRowUpper) {
-				// both set
-				rowIndices.add(incrementByOne(rowLower.expr, ctx));
-				rowIndices.add(rowUpper.expr);
-			}
-			else if(isRowLower && !isRowUpper) {
-				// only row set
-				rowIndices.add(incrementByOne(rowLower.expr, ctx));
-			}
-			else {
-				notifyErrorListeners("incorrect index expression for row", ctx.start);
-				return;
-			}
-
-			if(!isColLower && !isColUpper) {
-				// both not set
-				colIndices.add(null); colIndices.add(null);
-			}
-			else if(isColLower && isColUpper) {
-				colIndices.add(incrementByOne(colLower.expr, ctx));
-				colIndices.add(colUpper.expr);
-			}
-			else if(isColLower && !isColUpper) {
-				colIndices.add(incrementByOne(colLower.expr, ctx));
-			}
-			else {
-				notifyErrorListeners("incorrect index expression for column", ctx.start);
-				return;
-			}
-			exprList.add(rowIndices);
-			exprList.add(colIndices);
-			((IndexedIdentifier) dataInfo.expr).setIndices(exprList);
-		}
-		catch(Exception e) {
-			notifyErrorListeners("cannot set the indices", ctx.start);
-			return;
-		}
-	}
-
-	private Expression incrementByOne(Expression expr, ParserRuleContext ctx) {
-		// For maintaining semantic consistency, we have decided to keep 1-based indexing
-		// If in future, PyDML becomes more popular than DML, this can be switched.
-		return expr;
-	}
-
 	protected ConstIdentifier getConstIdFromString(String varValue, Token start) {
 
 		int linePosition = start.getLine();
