+++ src/main/java/org/apache/sysml/parser/pydml/PydmlSyntacticValidator.java
@@ -39,6 +39,7 @@ import org.apache.sysml.parser.BinaryExpression;
 import org.apache.sysml.parser.ConditionalPredicate;
 import org.apache.sysml.parser.DMLProgram;
 import org.apache.sysml.parser.DataIdentifier;
+import org.apache.sysml.parser.DoubleIdentifier;
 import org.apache.sysml.parser.Expression;
 import org.apache.sysml.parser.Expression.DataType;
 import org.apache.sysml.parser.Expression.ValueType;
@@ -49,6 +50,7 @@ import org.apache.sysml.parser.FunctionStatement;
 import org.apache.sysml.parser.IfStatement;
 import org.apache.sysml.parser.ImportStatement;
 import org.apache.sysml.parser.IntIdentifier;
+import org.apache.sysml.parser.IndexedIdentifier;
 import org.apache.sysml.parser.IterablePredicate;
 import org.apache.sysml.parser.LanguageException;
 import org.apache.sysml.parser.ParForStatement;
@@ -311,21 +313,95 @@ public class PydmlSyntacticValidator extends CommonSyntacticValidator implements
 	}
 
 
+	/**
+	 * PyDML uses 0-based indexing, so we increment lower indices by 1
+	 * when translating to DML.
+	 *
+	 * @param ctx the parse tree
+	 */
 	@Override
 	public void exitIndexedExpression(IndexedExpressionContext ctx) {
 		boolean isRowLower = (ctx.rowLower != null && !ctx.rowLower.isEmpty() && (ctx.rowLower.info.expr != null));
 		boolean isRowUpper = (ctx.rowUpper != null && !ctx.rowUpper.isEmpty() && (ctx.rowUpper.info.expr != null));
 		boolean isColLower = (ctx.colLower != null && !ctx.colLower.isEmpty() && (ctx.colLower.info.expr != null));
 		boolean isColUpper = (ctx.colUpper != null && !ctx.colUpper.isEmpty() && (ctx.colUpper.info.expr != null));
-		String name = ctx.name.getText();
-		exitIndexedExpressionHelper(ctx, name, ctx.dataInfo,
-				isRowLower ? ctx.rowLower.info : null,
-				isRowUpper ? ctx.rowUpper.info : null,
-				isColLower ? ctx.colLower.info : null,
-				isColUpper ? ctx.colUpper.info : null);
+        ExpressionInfo rowLower = isRowLower ? ctx.rowLower.info : null;
+        ExpressionInfo rowUpper = isRowUpper ? ctx.rowUpper.info : null;
+        ExpressionInfo colLower = isColLower ? ctx.colLower.info : null;
+        ExpressionInfo colUpper = isColUpper ? ctx.colUpper.info : null;
+
+		ctx.dataInfo.expr = new IndexedIdentifier(ctx.name.getText(), false, false);
+		setFileLineColumn(ctx.dataInfo.expr, ctx);
+
+		try {
+			ArrayList< ArrayList<Expression> > exprList = new ArrayList< ArrayList<Expression> >();
+
+			ArrayList<Expression> rowIndices = new ArrayList<Expression>();
+			ArrayList<Expression> colIndices = new ArrayList<Expression>();
+
+
+			if(!isRowLower && !isRowUpper) {
+				// both not set
+				rowIndices.add(null); rowIndices.add(null);
+			}
+			else if(isRowLower && isRowUpper) {
+				// both set
+				rowIndices.add(incrementByOne(rowLower.expr, ctx));
+				rowIndices.add(rowUpper.expr);
+			}
+			else if(isRowLower && !isRowUpper) {
+				// only row set
+				rowIndices.add(incrementByOne(rowLower.expr, ctx));
+			}
+			else {
+				notifyErrorListeners("incorrect index expression for row", ctx.start);
+				return;
+			}
+
+			if(!isColLower && !isColUpper) {
+				// both not set
+				colIndices.add(null); colIndices.add(null);
+			}
+			else if(isColLower && isColUpper) {
+				colIndices.add(incrementByOne(colLower.expr, ctx));
+				colIndices.add(colUpper.expr);
+			}
+			else if(isColLower && !isColUpper) {
+				colIndices.add(incrementByOne(colLower.expr, ctx));
+			}
+			else {
+				notifyErrorListeners("incorrect index expression for column", ctx.start);
+				return;
+			}
+			exprList.add(rowIndices);
+			exprList.add(colIndices);
+			((IndexedIdentifier) ctx.dataInfo.expr).setIndices(exprList);
+		}
+		catch(Exception e) {
+			notifyErrorListeners("cannot set the indices", ctx.start);
+			return;
+		}
 	}
 
-
+	/**
+	 * Increment lower indices by 1 when translating from PyDML
+	 * (0-based indexing) to DML (1-based indexing).
+	 *
+	 * @param expr
+	 * @param ctx
+	 * @return
+	 */
+	private Expression incrementByOne(Expression expr, ParserRuleContext ctx) {
+		// Addition and subtraction operator same as DML
+		Expression.BinaryOp bop = Expression.getBinaryOp("+");
+		Expression retVal = new BinaryExpression(bop);
+		((BinaryExpression)retVal).setLeft(expr);
+		int line = ctx.start.getLine();
+		int col = ctx.start.getCharPositionInLine();
+		((BinaryExpression)retVal).setRight(new DoubleIdentifier(1.0, currentFile, line, col, line, col));
+		setFileLineColumn(retVal, ctx);
+		return retVal;
+	}
 
 	// -----------------------------------------------------------------
 	//          Command line parameters (begin with a '$')
